ml-phase2/requirements.txt
# 机器学习模型依赖包
lightgbm==4.1.0
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pyarrow==14.0.1
optuna==3.5.0
shap==0.44.1